# docs/swagger.yaml
basePath: /portfolios
definitions:
  portfolios.AfterQuotePortfolio:
    properties:
      chain_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      tokens:
        items:
          $ref: '#/definitions/token.QuoteToken'
        type: array
    type: object
  portfolios.AfterSwapPortfolio:
    properties:
      chain_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      tokens:
        items:
          $ref: '#/definitions/token.SwapWrapperForToken'
        type: array
    type: object
  portfolios.InputPortfolio:
    properties:
      chain_id:
        type: integer
      name:
        type: string
      tokens:
        items:
          $ref: '#/definitions/token.InputToken'
        type: array
    type: object
  portfolios.ProportionsResponsePortfolio:
    properties:
      chain_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      tokensProportions:
        items:
          $ref: '#/definitions/token.ProportionsToken'
        type: array
    type: object
  portfolios.ResponsePortfolio:
    properties:
      chain_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      tokens:
        items:
          $ref: '#/definitions/token.PortfoliosToken'
        type: array
    type: object
  token.AfterSwapToken:
    properties:
      data:
        type: string
      from:
        type: string
      gas:
        type: integer
      gasPrice:
        type: string
      to:
        type: string
      value:
        type: string
    type: object
  token.InputToken:
    properties:
      address:
        type: string
      amount:
        type: string
    type: object
  token.InputTokens:
    properties:
      tokens:
        items:
          $ref: '#/definitions/token.InputToken'
        type: array
    type: object
  token.PortfoliosToken:
    properties:
      address:
        type: string
      amount:
        type: string
      decimals:
        type: integer
      ticker:
        type: string
    type: object
  token.ProportionsToken:
    properties:
      address:
        type: string
      proportion:
        type: number
      ticker:
        type: string
    type: object
  token.QuoteToken:
    properties:
      address:
        type: string
      estimated_gas:
        type: integer
      final_price:
        type: string
      native_price:
        type: string
      ticker:
        type: string
    type: object
  token.SwapWrapperForToken:
    properties:
      address:
        type: string
      ticker:
        type: string
      tx:
        $ref: '#/definitions/token.AfterSwapToken'
    type: object
host: localhost:8080
info:
  contact:
    email: KlenoviySir@yandex.ru
    name: API developer
    url: https://t.me/KlenoviySIr
  description: This is a portfolio API server. There are endpoints for creating, updating,
    deleting, getting portfolios and making swaps and quotes with them.
  title: Portfolio API
  version: "1.0"
paths:
  /:
    get:
      consumes:
      - '*/*'
      description: Get all portfolios in DB with their tokens with amount
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/portfolios.ResponsePortfolio'
            type: array
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get all portfolios in DB
      tags:
      - portfolio
    post:
      consumes:
      - application/json
      description: Create new portfolio with tokens
      parameters:
      - description: Portfolio
        in: body
        name: portfolio
        required: true
        schema:
          $ref: '#/definitions/portfolios.InputPortfolio'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/portfolios.ProportionsResponsePortfolio'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create new portfolio
      tags:
      - portfolio
  /{id}:
    delete:
      consumes:
      - '*/*'
      description: Delete portfolio with its tokens by id
      parameters:
      - description: Portfolio id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Delete portfolio by id
      tags:
      - portfolio
    get:
      consumes:
      - '*/*'
      description: Get portfolio by id with amount of tokens
      parameters:
      - description: Portfolio id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/portfolios.ResponsePortfolio'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get portfolio by id
      tags:
      - portfolio
    put:
      consumes:
      - application/json
      description: Update whole portfolio by id
      parameters:
      - description: Portfolio id
        in: path
        name: id
        required: true
        type: integer
      - description: Portfolio
        in: body
        name: portfolio
        required: true
        schema:
          $ref: '#/definitions/portfolios.InputPortfolio'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/portfolios.ProportionsResponsePortfolio'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Update portfolio
      tags:
      - portfolio
  /{id}/count:
    get:
      consumes:
      - '*/*'
      description: Get portfolio with pre-calculated tokens values for transaction
        and gas by portfolio id and token with amount from which will transaction
        be made
      parameters:
      - description: Portfolio id
        in: path
        name: id
        required: true
        type: integer
      - description: Amount of token
        in: query
        name: amount
        required: true
        type: string
      - description: Contract address of token
        in: query
        name: contractAddress
        required: true
        type: string
      - description: Gas price for transaction to count estimated gas of this chain
        in: query
        name: gasPrice
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/portfolios.AfterQuotePortfolio'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get portfolio after quote
      tags:
      - portfolio
  /{id}/proportions:
    get:
      consumes:
      - '*/*'
      description: Get portfolio with tokens represented with proportions by id
      parameters:
      - description: Portfolio id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/portfolios.ProportionsResponsePortfolio'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get portfolio proportions
      tags:
      - portfolio
  /{id}/tokens:
    patch:
      consumes:
      - application/json
      description: Add new tokens to portfolio by id
      parameters:
      - description: Portfolio id
        in: path
        name: id
        required: true
        type: integer
      - description: Tokens
        in: body
        name: tokens
        required: true
        schema:
          $ref: '#/definitions/token.InputTokens'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/portfolios.ProportionsResponsePortfolio'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Add new tokens to portfolio
      tags:
      - portfolio
  /{id}/transactions:
    get:
      consumes:
      - '*/*'
      description: Get transactions data for metamask for current portfolio with amount
        of token and wallet address
      parameters:
      - description: Portfolio id
        in: path
        name: id
        required: true
        type: integer
      - description: Amount of token
        in: query
        name: amount
        required: true
        type: string
      - description: Contract address of token
        in: query
        name: fromTokenAddress
        required: true
        type: string
      - description: User's wallet address
        in: query
        name: fromAddress
        required: true
        type: string
      - description: Slippage for transaction
        in: query
        name: slippage
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/portfolios.AfterSwapPortfolio'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get transactions
      tags:
      - portfolio
schemes:
- http
swagger: "2.0"